Move Task and Workflow into wf_data.py
